Rename component to NodeNetworkTab.
