--- ui/src/app/base/components/NetworkTab/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NetworkTab";
