+++ ui/src/app/base/components/NodeNetworkTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./NodeNetworkTab";
